update class name & other code review changes
