--- app/models/grda_warehouse/cohort_column_type.rb
@@ -1,310 +1,0 @@
-# frozen_string_literal: true
-
-module GrdaWarehouse
-  class CohortColumnType < GrdaWarehouseBase
-    validates :class_name, presence: true, uniqueness: true
-
-    scope :active, -> { where(active: true) }
-
-    def activate
-      update(active: true)
-    end
-
-    def deactivate
-      update(active: false)
-      remove_from_cohorts
-    end
-
-    def remove_from_cohorts
-      GrdaWarehouse::Cohort.all.each do |cohort|
-        cohort.update!(column_state: cohort.column_state.reject { |col| col.cohort_column_type.class_name == class_name })
-      end
-    end
-
-    def self.maintain!
-      known_cohort_column_types.each do |class_name|
-        column_type = find_by(class_name: class_name)
-        create!(class_name: class_name, active: true) unless column_type
-      end
-    end
-
-    def self.known_cohort_column_types
-      [
-        'CohortColumns::LastName',
-        'CohortColumns::FirstName',
-        'CohortColumns::Rank',
-        'CohortColumns::Age',
-        'CohortColumns::Dob',
-        'CohortColumns::Gender',
-        'CohortColumns::Ssn',
-        'CohortColumns::ClientId',
-        'CohortColumns::CalculatedDaysHomeless',
-        'CohortColumns::AdjustedDaysHomeless',
-        'CohortColumns::AdjustedDaysHomelessLastThreeYears',
-        'CohortColumns::AdjustedDaysLiterallyHomelessLastThreeYears',
-        'CohortColumns::DaysHomelessPlusOverrides',
-        'CohortColumns::FirstDateHomeless',
-        'CohortColumns::Chronic',
-        'CohortColumns::Agency',
-        'CohortColumns::CaseManager',
-        'CohortColumns::HousingManager',
-        'CohortColumns::HousingSearchAgency',
-        'CohortColumns::HousingOpportunity',
-        'CohortColumns::LegalBarriers',
-        'CohortColumns::CriminalRecordStatus',
-        'CohortColumns::DocumentReady',
-        'CohortColumns::SifEligible',
-        'CohortColumns::SensoryImpaired',
-        'CohortColumns::HousedDate',
-        'CohortColumns::Destination',
-        'CohortColumns::SubPopulation',
-        'CohortColumns::IndividualInMostRecentEnrollment',
-        'CohortColumns::StFrancisHouse',
-        'CohortColumns::LastGroupReviewDate',
-        'CohortColumns::LastDateApproached',
-        'CohortColumns::PreContemplativeLastDateApproached',
-        'CohortColumns::HousingTrackSuggested',
-        'CohortColumns::PrimaryHousingTrackSuggested',
-        'CohortColumns::HousingTrackEnrolled',
-        'CohortColumns::VaEligible',
-        'CohortColumns::VashEligible',
-        'CohortColumns::Chapter115',
-        'CohortColumns::Veteran',
-        'CohortColumns::ClientNotes',
-        'CohortColumns::Notes',
-        'CohortColumns::VispdatScore',
-        'CohortColumns::VispdatPriorityScore',
-        'CohortColumns::HousingNavigator',
-        'CohortColumns::LocationType',
-        'CohortColumns::Location',
-        'CohortColumns::LastContactLocation',
-        'CohortColumns::Status',
-        'CohortColumns::SsvfEligible',
-        'CohortColumns::VetSquaresConfirmed',
-        'CohortColumns::MissingDocuments',
-        'CohortColumns::Provider',
-        'CohortColumns::NextStep',
-        'CohortColumns::HousingPlan',
-        'CohortColumns::DateDocumentReady',
-        'CohortColumns::DaysHomelessLastThreeYears',
-        'CohortColumns::DaysLiterallyHomelessLastThreeYears',
-        'CohortColumns::ShelteredDaysHomelessLastThreeYears',
-        'CohortColumns::UnshelteredDaysHomelessLastThreeYears',
-        'CohortColumns::EnrolledHomelessShelter',
-        'CohortColumns::EnrolledHomelessUnsheltered',
-        'CohortColumns::EnrolledPermanentHousing',
-        'CohortColumns::RelatedUsers',
-        'CohortColumns::Active',
-        'CohortColumns::LastHomelessVisit',
-        'CohortColumns::OngoingEs',
-        'CohortColumns::OngoingSo',
-        'CohortColumns::OngoingSh',
-        'CohortColumns::OngoingTh',
-        'CohortColumns::OngoingRrh',
-        'CohortColumns::OngoingPsh',
-        'CohortColumns::NewLeaseReferral',
-        'CohortColumns::VulnerabilityRank',
-        'CohortColumns::ActiveCohorts',
-        'CohortColumns::DestinationFromHomelessness',
-        'CohortColumns::HmisDestination',
-        'CohortColumns::OpenEnrollments',
-        'CohortColumns::Ineligible',
-        'CohortColumns::ConsentConfirmed',
-        'CohortColumns::DisabilityVerificationDate',
-        'CohortColumns::AvailableForMatchingInCas',
-        'CohortColumns::DaysSinceCasMatch',
-        'CohortColumns::Sober',
-        'CohortColumns::OriginalChronic',
-        'CohortColumns::NotAVet',
-        'CohortColumns::EtoCoordinatedEntryAssessmentScore',
-        'CohortColumns::HouseholdMembers',
-        'CohortColumns::MinimumBedroomSize',
-        'CohortColumns::SpecialNeeds',
-        'CohortColumns::RrhDesired',
-        'CohortColumns::YouthRrhDesired',
-        'CohortColumns::RrhAssessmentContactInfo',
-        'CohortColumns::RrhSsvfEligible',
-        'CohortColumns::Reported',
-        'CohortColumns::Race',
-        'CohortColumns::Ethnicity',
-        'CohortColumns::Lgbtq',
-        'CohortColumns::LgbtqFromHmis',
-        'CohortColumns::SleepingLocation',
-        'CohortColumns::ExitDestination',
-        'CohortColumns::ActiveInCasMatch',
-        'CohortColumns::SchoolDistrict',
-        'CohortColumns::AssessmentScore',
-        'CohortColumns::PathwaysV3AssessmentDate',
-        'CohortColumns::TransferV3AssessmentDate',
-        'CohortColumns::VispdatScoreManual',
-        'CohortColumns::DaysOnCohort',
-        'CohortColumns::CasVashEligible',
-        'CohortColumns::DateAddedToCohort',
-        'CohortColumns::PreviousRemovalReason',
-        'CohortColumns::HealthPrioritized',
-        'CohortColumns::MostRecentDateToStreet',
-        'CohortColumns::DaysHomelessPathways',
-        'CohortColumns::MostRecentClsSheltered',
-        'CohortColumns::CeAssessmentDate',
-        'CohortColumns::CeAssessmentName',
-        'CohortColumns::CeAssessmentUser',
-        'CohortColumns::SourceClientPersonalIds',
-        'CohortColumns::MostRecentHouseholdType',
-        'CohortColumns::MostRecentSelfReportMonthsHomeless',
-        'CohortColumns::MostRecentPriorLivingSituation',
-        'CohortColumns::MostRecentCls',
-        'CohortColumns::VeteranStatusCalculated',
-        'CohortColumns::MostRecentDisablingCondition',
-        'CohortColumns::UserString1',
-        'CohortColumns::UserString2',
-        'CohortColumns::UserString3',
-        'CohortColumns::UserString4',
-        'CohortColumns::UserString5',
-        'CohortColumns::UserString6',
-        'CohortColumns::UserString7',
-        'CohortColumns::UserString8',
-        'CohortColumns::UserString9',
-        'CohortColumns::UserString10',
-        'CohortColumns::UserString11',
-        'CohortColumns::UserString12',
-        'CohortColumns::UserString13',
-        'CohortColumns::UserString14',
-        'CohortColumns::UserString15',
-        'CohortColumns::UserString16',
-        'CohortColumns::UserString17',
-        'CohortColumns::UserString18',
-        'CohortColumns::UserString19',
-        'CohortColumns::UserString20',
-        'CohortColumns::UserString21',
-        'CohortColumns::UserString22',
-        'CohortColumns::UserString23',
-        'CohortColumns::UserString24',
-        'CohortColumns::UserString25',
-        'CohortColumns::UserString26',
-        'CohortColumns::UserString27',
-        'CohortColumns::UserString28',
-        'CohortColumns::UserString29',
-        'CohortColumns::UserString30',
-        'CohortColumns::UserBoolean1',
-        'CohortColumns::UserBoolean2',
-        'CohortColumns::UserBoolean3',
-        'CohortColumns::UserBoolean4',
-        'CohortColumns::UserBoolean5',
-        'CohortColumns::UserBoolean6',
-        'CohortColumns::UserBoolean7',
-        'CohortColumns::UserBoolean8',
-        'CohortColumns::UserBoolean9',
-        'CohortColumns::UserBoolean10',
-        'CohortColumns::UserBoolean11',
-        'CohortColumns::UserBoolean12',
-        'CohortColumns::UserBoolean13',
-        'CohortColumns::UserBoolean14',
-        'CohortColumns::UserBoolean15',
-        'CohortColumns::UserBoolean16',
-        'CohortColumns::UserBoolean17',
-        'CohortColumns::UserBoolean18',
-        'CohortColumns::UserBoolean19',
-        'CohortColumns::UserBoolean20',
-        'CohortColumns::UserBoolean21',
-        'CohortColumns::UserBoolean22',
-        'CohortColumns::UserBoolean23',
-        'CohortColumns::UserBoolean24',
-        'CohortColumns::UserBoolean25',
-        'CohortColumns::UserBoolean26',
-        'CohortColumns::UserBoolean27',
-        'CohortColumns::UserBoolean28',
-        'CohortColumns::UserBoolean29',
-        'CohortColumns::UserBoolean30',
-        'CohortColumns::UserBoolean31',
-        'CohortColumns::UserBoolean32',
-        'CohortColumns::UserBoolean33',
-        'CohortColumns::UserBoolean34',
-        'CohortColumns::UserBoolean35',
-        'CohortColumns::UserBoolean36',
-        'CohortColumns::UserBoolean37',
-        'CohortColumns::UserBoolean38',
-        'CohortColumns::UserBoolean39',
-        'CohortColumns::UserBoolean40',
-        'CohortColumns::UserBoolean41',
-        'CohortColumns::UserBoolean42',
-        'CohortColumns::UserBoolean43',
-        'CohortColumns::UserBoolean44',
-        'CohortColumns::UserBoolean45',
-        'CohortColumns::UserBoolean46',
-        'CohortColumns::UserBoolean47',
-        'CohortColumns::UserBoolean48',
-        'CohortColumns::UserBoolean49',
-        'CohortColumns::UserSelect1',
-        'CohortColumns::UserSelect2',
-        'CohortColumns::UserSelect3',
-        'CohortColumns::UserSelect4',
-        'CohortColumns::UserSelect5',
-        'CohortColumns::UserSelect6',
-        'CohortColumns::UserSelect7',
-        'CohortColumns::UserSelect8',
-        'CohortColumns::UserSelect9',
-        'CohortColumns::UserSelect10',
-        'CohortColumns::UserSelect11',
-        'CohortColumns::UserSelect12',
-        'CohortColumns::UserSelect13',
-        'CohortColumns::UserSelect14',
-        'CohortColumns::UserSelect15',
-        'CohortColumns::UserSelect16',
-        'CohortColumns::UserSelect17',
-        'CohortColumns::UserSelect18',
-        'CohortColumns::UserSelect19',
-        'CohortColumns::UserSelect20',
-        'CohortColumns::UserSelect21',
-        'CohortColumns::UserSelect22',
-        'CohortColumns::UserSelect23',
-        'CohortColumns::UserSelect24',
-        'CohortColumns::UserSelect25',
-        'CohortColumns::UserSelect26',
-        'CohortColumns::UserSelect27',
-        'CohortColumns::UserSelect28',
-        'CohortColumns::UserSelect29',
-        'CohortColumns::UserSelect30',
-        'CohortColumns::UserDate1',
-        'CohortColumns::UserDate2',
-        'CohortColumns::UserDate3',
-        'CohortColumns::UserDate4',
-        'CohortColumns::UserDate5',
-        'CohortColumns::UserDate6',
-        'CohortColumns::UserDate7',
-        'CohortColumns::UserDate8',
-        'CohortColumns::UserDate9',
-        'CohortColumns::UserDate10',
-        'CohortColumns::UserDate11',
-        'CohortColumns::UserDate12',
-        'CohortColumns::UserDate13',
-        'CohortColumns::UserDate14',
-        'CohortColumns::UserDate15',
-        'CohortColumns::UserDate16',
-        'CohortColumns::UserDate17',
-        'CohortColumns::UserDate18',
-        'CohortColumns::UserDate19',
-        'CohortColumns::UserDate20',
-        'CohortColumns::UserDate21',
-        'CohortColumns::UserDate22',
-        'CohortColumns::UserDate23',
-        'CohortColumns::UserDate24',
-        'CohortColumns::UserDate25',
-        'CohortColumns::UserDate26',
-        'CohortColumns::UserDate27',
-        'CohortColumns::UserDate28',
-        'CohortColumns::UserDate29',
-        'CohortColumns::UserDate30',
-        'CohortColumns::UserNumeric1',
-        'CohortColumns::UserNumeric2',
-        'CohortColumns::UserNumeric3',
-        'CohortColumns::UserNumeric4',
-        'CohortColumns::UserNumeric5',
-        'CohortColumns::UserNumeric6',
-        'CohortColumns::UserNumeric7',
-        'CohortColumns::UserNumeric8',
-        'CohortColumns::UserNumeric9',
-        'CohortColumns::UserNumeric10',
-      ]
-    end
-  end
-end
