+++ db/warehouse/migrate/20250409145017_add_cohort_column_table.rb
@@ -2,7 +2,7 @@
 
 class AddCohortColumnTable < ActiveRecord::Migration[7.0]
   def change
-    create_table :cohort_column_types do |t|
+    create_table :cohort_columns do |t|
       t.string :class_name, unique: true
       t.boolean :active, default: true
     end
