+++ db/warehouse/migrate/20250409145018_populate_cohort_column_table.rb
@@ -2,10 +2,10 @@
 
 class PopulateCohortColumnTable < ActiveRecord::Migration[7.0]
   def up
-    GrdaWarehouse::CohortColumnType.maintain!
+    GrdaWarehouse::Cohorts::CohortColumn.maintain!
   end
 
   def down
-    GrdaWarehouse::CohortColumnType.destroy_all
+    GrdaWarehouse::Cohorts::CohortColumn.destroy_all
   end
 end
