+++ spec/factories/cohort_columns/adjusted_days_homeless.rb
@@ -2,6 +2,6 @@
 
 FactoryBot.define do
   factory :adjusted_days_homeless, class: 'CohortColumns::AdjustedDaysHomeless' do
-    cohort_column_type { GrdaWarehouse::CohortColumnType.find_by(class_name: 'CohortColumns::AdjustedDaysHomeless') }
+    cohort_column { GrdaWarehouse::Cohorts::CohortColumn.find_by(class_name: 'CohortColumns::AdjustedDaysHomeless') }
   end
 end
