+++ spec/factories/cohort_columns/date_added_to_cohort.rb
@@ -2,6 +2,6 @@
 
 FactoryBot.define do
   factory :date_added_to_cohort, class: 'CohortColumns::DateAddedToCohort' do
-    cohort_column_type { GrdaWarehouse::CohortColumnType.find_by(class_name: 'CohortColumns::DateAddedToCohort') }
+    cohort_column { GrdaWarehouse::Cohorts::CohortColumn.find_by(class_name: 'CohortColumns::DateAddedToCohort') }
   end
 end
