+++ spec/factories/cohort_columns/rank.rb
@@ -2,6 +2,6 @@
 
 FactoryBot.define do
   factory :rank, class: 'CohortColumns::Rank' do
-    cohort_column_type { GrdaWarehouse::CohortColumnType.find_by(class_name: 'CohortColumns::Rank') }
+    cohort_column { GrdaWarehouse::Cohorts::CohortColumn.find_by(class_name: 'CohortColumns::Rank') }
   end
 end
