+++ spec/factories/cohort_columns/user_string.rb
@@ -4,7 +4,7 @@ FactoryBot.define do
   1..30.times do |i|
     name = "user_string_cohort_column_#{i}"
     factory name.to_sym, class: "CohortColumns::UserString#{i}" do
-      cohort_column_type { GrdaWarehouse::CohortColumnType.find_by(class_name: "CohortColumns::UserString#{i}") }
+      cohort_column { GrdaWarehouse::Cohorts::CohortColumn.find_by(class_name: "CohortColumns::UserString#{i}") }
     end
   end
 end
