+++ spec/factories/grda_warehouse/cohort_column_type.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  factory :cohort_column_type, class: 'GrdaWarehouse::CohortColumnType' do
+  factory :cohort_column, class: 'GrdaWarehouse::Cohorts::CohortColumn' do
     sequence(:class_name) { |n| "CohortColumns::UserString#{n}" }
     active { true }
   end
