+++ spec/models/cohort_columns/date_added_to_cohort_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe CohortColumns::DateAddedToCohort, type: :model do
 
   before(:all) do
     GrdaWarehouse::ServiceHistoryServiceMaterialized.rebuild!
-    GrdaWarehouse::CohortColumnType.maintain!
+    GrdaWarehouse::Cohorts::CohortColumn.maintain!
   end
 
   it 'Sets the date added to the cohort client' do
